backend/app/requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.30.0
babel==2.17.0
bcrypt==5.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
et_xmlfile==2.0.0
fastapi==0.104.1
fastapi-admin==1.0.4
GeoAlchemy2==0.14.2
greenlet==3.2.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.5
packaging==25.0
pandas==2.1.3
pandas-stubs==2.3.2.250827
pendulum==3.1.0
psycopg2-binary==2.9.9
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
pypika-tortoise==0.6.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
ruff==0.13.2
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.21.0
SQLAlchemy==2.0.23
starlette==0.27.0
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
WTForms==3.1.2
